Prevent CPU OOM when seq_len very large (e.g. 128k)
attention_mask is not used by the Mamba model. When seq_len is very
large, attention_mask will be seq_len^2, which causes the CPU to OOM
(for seq_len=128k). This commit changes the unused attention_mask to
contain only one element, passes None for attention_mask into the model,
and adds an assertion to the model so that someone does not try to use
it with non-None attention_mask in another context (such as while
running an evaluation).